move everthing discord related to its own crates and split everything into multiple files
