+++ crates/calendarbot/src/discord/local_cache.rs
@@ -1,0 +1,27 @@
+use std::sync::Arc;
+
+use poise::serenity_prelude as serenity;
+
+#[derive(Clone)]
+pub struct LocalCache {
+    pub(crate) cache: Arc<serenity::Cache>,
+    pub(crate) client: Arc<serenity::Http>,
+}
+impl LocalCache {
+    pub(crate) fn new(client: Arc<serenity::Http>) -> Self {
+        Self {
+            cache: Arc::new(serenity::Cache::default()),
+            client,
+        }
+    }
+}
+
+impl serenity::CacheHttp for LocalCache {
+    fn http(&self) -> &serenity::Http {
+        &self.client
+    }
+
+    fn cache(&self) -> Option<&Arc<serenity::Cache>> {
+        Some(&self.cache)
+    }
+}
