+++ crates/calendarbot/src/main.rs
@@ -5,7 +5,6 @@ This program comes with ABSOLUTELY NO WARRANTY; for details type `show w'.
 This is free software, and you are welcome to redistribute it
  */
 mod calendar;
-mod commands;
 mod discord;
 pub mod models;
 pub mod schema;
